refactor(tests): give the whole-tree sweeps their own `tests/guards` layer
