+++ pyproject.toml
@@ -100,6 +100,7 @@ asyncio_default_fixture_loop_scope = "function"
 timeout = 120
 markers = [
     "unit: fast, no network or filesystem beyond the repo",
+    "guard: asserts a property of the repository rather than the behaviour of one module",
     "integration: opens a real socket or talks to a real service",
 ]
