+++ tests/guards/name_resolution.py
